refactor(coreextended): fix coreextented -> coreextended typo (breaking)
Rename the misspelled package: the directory lib/coreextented and the
Go package name were both 'coreextented' (missing a 'd'), while the
files inside (coreextended.go, header-coreextended.lisp) and the child
package nscoreextended were already spelled correctly.

Breaking change with no back-compat shim: embedders importing
github.com/jig/lisp/lib/coreextented (or its nscoreextended subpackage)
must update the path to .../lib/coreextended when moving to jig/lisp
0.3. The package identifier used in code (coreextended.HeaderCoreExtended)
changes the same way.
