+++ macro_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/jig/lisp/env"
 	"github.com/jig/lisp/lib/concurrent"
 	"github.com/jig/lisp/lib/core"
-	"github.com/jig/lisp/lib/coreextented"
+	"github.com/jig/lisp/lib/coreextended"
 	"github.com/jig/lisp/types"
 )
 
@@ -547,7 +547,7 @@ func TestMacro(t *testing.T) {
 	if _, err := REPL(ctx, repl_env, core.HeaderLoadFile(), types.NewCursorFile(reflect.TypeOf(_here_{}).PkgPath())); err != nil {
 		t.Fatal(err)
 	}
-	if _, err := REPL(ctx, repl_env, coreextented.HeaderCoreExtended(), types.NewCursorFile(reflect.TypeOf(_here_{}).PkgPath())); err != nil {
+	if _, err := REPL(ctx, repl_env, coreextended.HeaderCoreExtended(), types.NewCursorFile(reflect.TypeOf(_here_{}).PkgPath())); err != nil {
 		t.Fatal(err)
 	}
 	if _, err := REPL(ctx, repl_env, concurrent.HeaderConcurrent(), types.NewCursorFile(reflect.TypeOf(_here_{}).PkgPath())); err != nil {
